refactor(ui): merge Pill into Eyebrow and migrate all consumers
