+++ packages/ui/components/ui/Eyebrow/index.tsx
@@ -1,23 +1,33 @@
 import { cn } from "../../../utils";
 
+type Tone = "default" | "primary" | "muted" | "accent" | "outline";
+
 interface Props {
   children: React.ReactNode;
-  /** "dot" = orange circle prefix; "dash" = em-dash prefix; "none" = no prefix */
+  /** "dot" = circle prefix (adapts to chip text color); "dash" = em-dash prefix; "none" = no prefix */
   prefix?: "dot" | "dash" | "none";
-  tone?: "default" | "primary" | "muted";
+  tone?: Tone;
+  size?: "sm" | "md";
   className?: string;
 }
 
-const toneMap = {
-  default: "text-foreground",
-  primary: "text-primary",
-  muted: "text-muted-foreground",
-} as const;
+const toneMap: Record<Tone, string> = {
+  default: "bg-foreground text-background",
+  primary: "bg-primary text-primary-foreground",
+  muted: "text-foreground",
+  accent: "bg-accent text-accent-foreground",
+  outline: "border border-foreground text-foreground",
+};
 
-export function Eyebrow({ children, prefix = "none", tone = "primary", className }: Props) {
+const sizeMap = {
+  sm: "px-2.5 py-1 text-[10px]",
+  md: "px-3.5 py-[7px] text-[0.72rem] font-semibold",
+};
+
+export function Eyebrow({ children, prefix = "none", tone = "primary", size = "md", className }: Props) {
   return (
-    <span className={cn("inline-flex items-center gap-2 font-mono text-[0.72rem] font-medium uppercase tracking-[0.16em]", toneMap[tone], className)}>
-      {prefix === "dot" && <span aria-hidden className="inline-block size-1.5 rounded-pill bg-primary" />}
+    <span className={cn("inline-flex w-fit items-center gap-1.5 rounded-pill font-mono uppercase tracking-[0.16em] leading-none whitespace-nowrap", toneMap[tone], sizeMap[size], className)}>
+      {prefix === "dot" && <span aria-hidden className="inline-block size-1.5 rounded-pill bg-current" />}
       {prefix === "dash" && <span aria-hidden>—</span>}
       <span>{children}</span>
     </span>
