+++ packages/ui/components/ui/SectionHeader/index.tsx
@@ -1,6 +1,6 @@
 import { cn } from "../../../utils";
 import { DisplayHeading } from "../DisplayHeading";
-import { Pill } from "../Pill";
+import { Eyebrow } from "../Eyebrow";
 
 interface SectionHeaderProps {
   badge?: string | null;
@@ -15,7 +15,11 @@ interface SectionHeaderProps {
 export function SectionHeader({ badge, heading, headingAs = "h2", headingSize = "display-2", lead, align = "left", className }: SectionHeaderProps) {
   return (
     <div className={cn("flex max-w-[720px] flex-col gap-5", align === "center" && "items-center text-center", className)}>
-      {badge && <Pill tone="accent">{badge}</Pill>}
+      {badge && (
+        <Eyebrow prefix="dot" tone="accent">
+          {badge}
+        </Eyebrow>
+      )}
       <DisplayHeading as={headingAs} size={headingSize} text={heading} />
       {lead && <p className="text-lead text-muted-foreground">{lead}</p>}
     </div>
